refactor: move rate counter to pkg
