+++ webook/pkg/rate_counter/sliding_window_test.go
@@ -1,4 +1,4 @@
-package fault_detect
+package rate_counter
 
 import (
 	"context"
